refactor: major refactor - use type alaising, reorganize files, decouple shared objects
