+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node dist/app.js",
-    "build": "npx tsc",
+    "build": "npx tsc && npx tsc-alias",
     "dev": "npx nodemon -w *.ts -e ts -x ts-node --files -H -T ./src/app.ts"
   },
   "nodemonConfig": {
@@ -20,6 +20,8 @@
     "@types/node-cron": "^3.0.10",
     "nodemon": "^3.0.1",
     "ts-node": "^10.9.2",
+    "tsc-alias": "^1.8.8",
+    "tsconfig-paths": "^4.2.0",
     "typescript": "^5.2.2"
   },
   "dependencies": {
