+++ src/createBot.ts
@@ -1,44 +1,35 @@
-import { Telegraf } from "telegraf";
-import "dotenv/config";
-import start from "./handlers/start";
-import help from "./handlers/help";
-import cancel from "./handlers/cancel";
-import result from "./handlers/result";
-import calendar from "./handlers/calendar";
-import search from "./handlers/search";
-import code from "./handlers/code";
-import changeFilter from "./handlers/changeFilter";
-import notifications from "./handlers/notifications";
-import subscribe from "./handlers/subscribe";
-import unsubscribe from "./handlers/unsubscribe";
-import timetable from "./handlers/timetable";
-import defaultHandler from "./handlers/defaultHandler";
+import bot from "./bot";
+import start from "@handlers/start";
+import help from "@handlers/help";
+import cancel from "@handlers/cancel";
+import result from "@handlers/result";
+import calendar from "@handlers/calendar";
+import search from "@handlers/search";
+import code from "@handlers/code";
+import changeFilter from "@handlers/changeFilter";
+import notifications from "@handlers/notifications";
+import subscribe from "@handlers/subscribe";
+import unsubscribe from "@handlers/unsubscribe";
+import timetable from "@handlers/timetable";
+import defaultHandler from "@handlers/defaultHandler";
 import { Scenes, session } from "telegraf";
-import resultWizard from "./wizards/resultWizard";
-import academicCalendarWizard from "./wizards/calendarWizard";
-import announcementWizard from "./wizards/announcementWizard";
-import timetableWizard from "./wizards/timeTableWizard";
-import { CustomContext } from "./types/customContext.type";
-import availableCommands from "./constants/availableCommands";
-import loggingMiddleware from "./middlewares/loggingMiddleware";
-import throttler from "./middlewares/throttler";
+import resultWizard from "@wizards/resultWizard";
+import academicCalendarWizard from "@wizards/calendarWizard";
+import announcementWizard from "@wizards/announcementWizard";
+import timetableWizard from "@wizards/timeTableWizard";
+import { CustomContext } from "@/types/customContext.type";
+import availableCommands from "@constants/availableCommands";
+import loggingMiddleware from "@middlewares/loggingMiddleware";
+import throttler from "@middlewares/throttler";
 import {
   inlineQueryResultHandler,
   searchInlineQueryHandler,
-} from "./handlers/searchInlineQueryHandler";
-import filterCallbackHandler from "./handlers/filterCallbackHandler";
+} from "@handlers/searchInlineQueryHandler";
+import filterCallbackHandler from "@handlers/filterCallbackHandler";
 
-function createBot() {
-  // Create the bot
-  const opts = {
-    // During result publish times, KTU servers will be slow to respond, this makes the API requests to be slower
-    // Axios timeout is set to 25 seconds, but telegraf will timeout before that
-    // This creates request duplication. Hence we need to increase the telegraf timeout
-    handlerTimeout: 1000 * 30,
-  };
-  const bot = new Telegraf<CustomContext>(process.env.BOT_TOKEN!, opts);
-
-  // Create wizards
+// Attach all middlewares to the bot
+function attachMiddlewares() {
+  // Create wizard
   const stage = new Scenes.Stage<CustomContext>([
     resultWizard,
     announcementWizard,
@@ -51,12 +42,15 @@ function createBot() {
   bot.use(loggingMiddleware);
   bot.use(session());
   bot.use(stage.middleware());
+}
 
-  // Register available commands
+// Set the available commands for the bot
+function setCommands() {
   bot.telegram.setMyCommands(availableCommands);
+}
 
-  // Register handlers
-  bot.start(async (ctx) => await start(ctx));
+// Attach all command handlers to the bot
+function attachCommands() {
   bot.command("help", async (ctx) => await help(ctx));
   bot.command("search", async (ctx) => await search(ctx));
   bot.command("result", async (ctx) => await result(ctx));
@@ -68,20 +62,25 @@ function createBot() {
   bot.command("timetable", async (ctx) => await timetable(ctx));
   bot.command("calendar", async (ctx) => await calendar(ctx));
   bot.command("changefilter", async (ctx) => await changeFilter(ctx));
+}
+
+// Attach all other event listeners to the bot
+function attachListeners() {
+  bot.start(async (ctx) => await start(ctx));
   bot.action(/filter_*/, async (ctx) => await filterCallbackHandler(ctx));
   bot.on("inline_query", async (ctx) => await searchInlineQueryHandler(ctx));
   bot.on("chosen_inline_result", async (ctx) => {
     await inlineQueryResultHandler(ctx.chosenInlineResult, bot);
   });
   bot.on("message", async (ctx) => await defaultHandler(ctx));
+}
 
-  // The top level error handler
-  // this will catch any errors that may happen
-  bot.catch((error) => {
-    console.log("telegraf error", error);
-  });
-
-  return bot;
+// Attach all commands, middlewares and listeners to the bot
+async function createBot() {
+  attachMiddlewares();
+  setCommands();
+  attachCommands();
+  attachListeners();
 }
 
 export default createBot;
