+++ src/handlers/calendar.ts
@@ -1,4 +1,4 @@
-import { CustomContext } from "../types/customContext.type";
+import { CustomContext } from "@/types/customContext.type";
 
 async function calendar(ctx: CustomContext) {
   await ctx.scene.enter("academic-calendar-wizard");
