+++ src/handlers/changeFilter.ts
@@ -1,9 +1,9 @@
 // Handler to change notification filter of user
-import { CustomContext } from "../types/customContext.type";
-import { FILTERS } from "../constants/constants";
+import { CustomContext } from "@/types/customContext.type";
+import { FILTERS } from "@/constants/constants";
 import { Markup } from "telegraf";
-import deleteMessage from "../utils/deleteMessage";
-import db from "../db/initDb";
+import deleteMessage from "@/utils/deleteMessage";
+import db from "@/db/initDb";
 
 async function changeFilter(ctx: CustomContext) {
   const waitingMsg = await ctx.reply("Please wait...");
