+++ src/handlers/subscribe.ts
@@ -1,8 +1,8 @@
-import { CustomContext } from "../types/customContext.type";
-import { FILTERS } from "../constants/constants";
+import { CustomContext } from "@/types/customContext.type";
+import { FILTERS } from "@/constants/constants";
 import { Markup } from "telegraf";
-import deleteMessage from "../utils/deleteMessage";
-import db from "../db/initDb";
+import deleteMessage from "@/utils/deleteMessage";
+import db from "@/db/initDb";
 
 async function subscribe(ctx: CustomContext) {
   const waitingMsg = await ctx.reply("Please wait...");
