+++ src/handlers/timetable.ts
@@ -1,4 +1,4 @@
-import { CustomContext } from "../types/customContext.type";
+import { CustomContext } from "@/types/customContext.type";
 
 async function timetable(ctx: CustomContext) {
   await ctx.scene.enter("timetable-wizard");
