+++ src/services/fetchAcademicCalendars.ts
@@ -1,8 +1,8 @@
-import { axios } from "../config/axiosConfig";
-import { ACADEMIC_CALENDAR_URL } from "../constants/constants";
-import { AcademicCalendar } from "../types/types";
-import ServerError from "../errors/ServerError";
-import formatDate from "../utils/formatDate";
+import { axios } from "@/config/axiosConfig";
+import { ACADEMIC_CALENDAR_URL } from "@/constants/constants";
+import { AcademicCalendar } from "@/types/types";
+import ServerError from "@/errors/ServerError";
+import formatDate from "@/utils/formatDate";
 
 async function fetchAcademicCalendars(
   pageNumber: number,
